Check inn to "save" various progress
Updating how drag and drop and upgrading stuff with tech is working. Now ship can be upgraded with upgrades without needing a free tech slot.
Also implemented selling (which was missing...), but when trying to make a resonable indication for selling I have a bug. I try to make the object change when selling, instead of making the slot change.