Fix: needed to use ref on DataStreamReader
otherwise the deserialization ended up incorrect.
As a side-note: tested with 5 ships with 5 tech each, msg size was 564B (current maximum)
This meeans we can expand quite a bit and be inside default max packet size of 1500, and the "normal" size of 1470
There are also plenty of ways I could reduce data-size, as there are plenty of "unused" bits that are now sent
E.g. many fields can probably be 8bit instead of 16, and for tech I can reuse parameters that are "type specific" so that I dont send lots of data that is not relevant for the given equipment